rework for splitting code in multiple files
